Move all ships logic to one file and move that file to be in the GameBoard folder. Remove the now unecessary using statement from GameBoard class.
