+++ CipherSink/Models/GameBoard/Ships.cs
@@ -1,6 +1,4 @@
-﻿using CipherSink.Models.GameBoard;
-
-namespace CipherSink.Models.Ships;
+﻿namespace CipherSink.Models.GameBoard;
 
 /// <summary>
 /// This class represents a ship on the game board.
@@ -86,4 +84,30 @@ internal abstract class Ship
 
         Positions = positions;
     }
+}
+
+
+internal class Carrier : Ship
+{
+    public Carrier() : base("Carrier", 5, OccupationType.Carrier) { }
+}
+
+internal class Battleship : Ship
+{
+    public Battleship() : base("Battleship", 4, OccupationType.Battleship) { }
+}
+
+internal class Cruiser : Ship
+{
+    public Cruiser() : base("Cruiser", 3, OccupationType.Cruiser) { }
+}
+
+internal class Submarine : Ship
+{
+    public Submarine() : base("Submarine", 3, OccupationType.Submarine) { }
+}
+
+internal class Destroyer : Ship
+{
+    public Destroyer() : base("Destroyer", 2, OccupationType.Destroyer) { }
 }
